split cv2.cpp: highgui, more utils
