+++ modules/python/src2/cv2.cpp
@@ -6,13 +6,12 @@
 
 #include "pyopencv_generated_include.h"
 #include "opencv2/core/types_c.h"
-#include <map>
-
-#include <type_traits>  // std::enable_if
 
 
 #include "cv2_util.hpp"
-
+#include "cv2_numpy.hpp"
+#include "cv2_convert.hpp"
+#include "cv2_highgui.hpp"
 
 using namespace cv;
 
@@ -50,234 +49,8 @@ typedef std::vector<std::vector<Point3f> > vector_vector_Point3f;
 typedef std::vector<std::vector<DMatch> > vector_vector_DMatch;
 typedef std::vector<std::vector<KeyPoint> > vector_vector_KeyPoint;
 
-#include "cv2_numpy.hpp"
+// enum { ARG_NONE = 0, ARG_MAT = 1, ARG_SCALAR = 2 };
 
-enum { ARG_NONE = 0, ARG_MAT = 1, ARG_SCALAR = 2 };
-
-#include "cv2_convert.hpp"
-
-
-
-static int OnError(int status, const char *func_name, const char *err_msg, const char *file_name, int line, void *userdata)
-{
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-
-    PyObject *on_error = (PyObject*)userdata;
-    PyObject *args = Py_BuildValue("isssi", status, func_name, err_msg, file_name, line);
-
-    PyObject *r = PyObject_Call(on_error, args, NULL);
-    if (r == NULL) {
-        PyErr_Print();
-    } else {
-        Py_DECREF(r);
-    }
-
-    Py_DECREF(args);
-    PyGILState_Release(gstate);
-
-    return 0; // The return value isn't used
-}
-
-static PyObject *pycvRedirectError(PyObject*, PyObject *args, PyObject *kw)
-{
-    const char *keywords[] = { "on_error", NULL };
-    PyObject *on_error;
-
-    if (!PyArg_ParseTupleAndKeywords(args, kw, "O", (char**)keywords, &on_error))
-        return NULL;
-
-    if ((on_error != Py_None) && !PyCallable_Check(on_error))  {
-        PyErr_SetString(PyExc_TypeError, "on_error must be callable");
-        return NULL;
-    }
-
-    // Keep track of the previous handler parameter, so we can decref it when no longer used
-    static PyObject* last_on_error = NULL;
-    if (last_on_error) {
-        Py_DECREF(last_on_error);
-        last_on_error = NULL;
-    }
-
-    if (on_error == Py_None) {
-        ERRWRAP2(redirectError(NULL));
-    } else {
-        last_on_error = on_error;
-        Py_INCREF(last_on_error);
-        ERRWRAP2(redirectError(OnError, last_on_error));
-    }
-    Py_RETURN_NONE;
-}
-
-static void OnMouse(int event, int x, int y, int flags, void* param)
-{
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-
-    PyObject *o = (PyObject*)param;
-    PyObject *args = Py_BuildValue("iiiiO", event, x, y, flags, PyTuple_GetItem(o, 1));
-
-    PyObject *r = PyObject_Call(PyTuple_GetItem(o, 0), args, NULL);
-    if (r == NULL)
-        PyErr_Print();
-    else
-        Py_DECREF(r);
-    Py_DECREF(args);
-    PyGILState_Release(gstate);
-}
-
-#ifdef HAVE_OPENCV_HIGHGUI
-static PyObject *pycvSetMouseCallback(PyObject*, PyObject *args, PyObject *kw)
-{
-    const char *keywords[] = { "window_name", "on_mouse", "param", NULL };
-    char* name;
-    PyObject *on_mouse;
-    PyObject *param = NULL;
-
-    if (!PyArg_ParseTupleAndKeywords(args, kw, "sO|O", (char**)keywords, &name, &on_mouse, &param))
-        return NULL;
-    if (!PyCallable_Check(on_mouse)) {
-        PyErr_SetString(PyExc_TypeError, "on_mouse must be callable");
-        return NULL;
-    }
-    if (param == NULL) {
-        param = Py_None;
-    }
-    PyObject* py_callback_info = Py_BuildValue("OO", on_mouse, param);
-    static std::map<std::string, PyObject*> registered_callbacks;
-    std::map<std::string, PyObject*>::iterator i = registered_callbacks.find(name);
-    if (i != registered_callbacks.end())
-    {
-        Py_DECREF(i->second);
-        i->second = py_callback_info;
-    }
-    else
-    {
-        registered_callbacks.insert(std::pair<std::string, PyObject*>(std::string(name), py_callback_info));
-    }
-    ERRWRAP2(setMouseCallback(name, OnMouse, py_callback_info));
-    Py_RETURN_NONE;
-}
-#endif
-
-static void OnChange(int pos, void *param)
-{
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-
-    PyObject *o = (PyObject*)param;
-    PyObject *args = Py_BuildValue("(i)", pos);
-    PyObject *r = PyObject_Call(PyTuple_GetItem(o, 0), args, NULL);
-    if (r == NULL)
-        PyErr_Print();
-    else
-        Py_DECREF(r);
-    Py_DECREF(args);
-    PyGILState_Release(gstate);
-}
-
-#ifdef HAVE_OPENCV_HIGHGUI
-// workaround for #20408, use nullptr, set value later
-static int _createTrackbar(const String &trackbar_name, const String &window_name, int value, int count,
-                    TrackbarCallback onChange, PyObject* py_callback_info)
-{
-    int n = createTrackbar(trackbar_name, window_name, NULL, count, onChange, py_callback_info);
-    setTrackbarPos(trackbar_name, window_name, value);
-    return n;
-}
-static PyObject *pycvCreateTrackbar(PyObject*, PyObject *args)
-{
-    PyObject *on_change;
-    char* trackbar_name;
-    char* window_name;
-    int value;
-    int count;
-
-    if (!PyArg_ParseTuple(args, "ssiiO", &trackbar_name, &window_name, &value, &count, &on_change))
-        return NULL;
-    if (!PyCallable_Check(on_change)) {
-        PyErr_SetString(PyExc_TypeError, "on_change must be callable");
-        return NULL;
-    }
-    PyObject* py_callback_info = Py_BuildValue("OO", on_change, Py_None);
-    std::string name = std::string(window_name) + ":" + std::string(trackbar_name);
-    static std::map<std::string, PyObject*> registered_callbacks;
-    std::map<std::string, PyObject*>::iterator i = registered_callbacks.find(name);
-    if (i != registered_callbacks.end())
-    {
-        Py_DECREF(i->second);
-        i->second = py_callback_info;
-    }
-    else
-    {
-        registered_callbacks.insert(std::pair<std::string, PyObject*>(name, py_callback_info));
-    }
-    ERRWRAP2(_createTrackbar(trackbar_name, window_name, value, count, OnChange, py_callback_info));
-    Py_RETURN_NONE;
-}
-
-static void OnButtonChange(int state, void *param)
-{
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-
-    PyObject *o = (PyObject*)param;
-    PyObject *args;
-    if(PyTuple_GetItem(o, 1) != NULL)
-    {
-        args = Py_BuildValue("(iO)", state, PyTuple_GetItem(o,1));
-    }
-    else
-    {
-        args = Py_BuildValue("(i)", state);
-    }
-
-    PyObject *r = PyObject_Call(PyTuple_GetItem(o, 0), args, NULL);
-    if (r == NULL)
-        PyErr_Print();
-    else
-        Py_DECREF(r);
-    Py_DECREF(args);
-    PyGILState_Release(gstate);
-}
-
-static PyObject *pycvCreateButton(PyObject*, PyObject *args, PyObject *kw)
-{
-    const char* keywords[] = {"buttonName", "onChange", "userData", "buttonType", "initialButtonState", NULL};
-    PyObject *on_change;
-    PyObject *userdata = NULL;
-    char* button_name;
-    int button_type = 0;
-    int initial_button_state = 0;
-
-    if (!PyArg_ParseTupleAndKeywords(args, kw, "sO|Oii", (char**)keywords, &button_name, &on_change, &userdata, &button_type, &initial_button_state))
-        return NULL;
-    if (!PyCallable_Check(on_change)) {
-        PyErr_SetString(PyExc_TypeError, "onChange must be callable");
-        return NULL;
-    }
-    if (userdata == NULL) {
-        userdata = Py_None;
-    }
-
-    PyObject* py_callback_info = Py_BuildValue("OO", on_change, userdata);
-    std::string name(button_name);
-
-    static std::map<std::string, PyObject*> registered_callbacks;
-    std::map<std::string, PyObject*>::iterator i = registered_callbacks.find(name);
-    if (i != registered_callbacks.end())
-    {
-        Py_DECREF(i->second);
-        i->second = py_callback_info;
-    }
-    else
-    {
-        registered_callbacks.insert(std::pair<std::string, PyObject*>(name, py_callback_info));
-    }
-    ERRWRAP2(createButton(button_name, OnButtonChange, py_callback_info, button_type, initial_button_state != 0));
-    Py_RETURN_NONE;
-}
-#endif
 
 ///////////////////////////////////////////////////////////////////////////////////////
